SAGE-001 Added landing page, sidebar, header
Added landing home page so users can see an initial call to action
and sign up.
Added header present on all pages, will have different options for
users signed in vs signed out. For now it is just a SignIn button.
Added sidebar for future navigation
