+++ frontend/src/components/Header.tsx
@@ -1,0 +1,32 @@
+export function Header() {
+
+    /*
+        TODO
+        Header has StockSage Name, Search, Profile, and Settings
+        TODO
+
+        TODO
+        Add Search, Profile, and Settings only when signed in
+        TODO
+    */
+    return (
+        <header className="border-b bg-card">
+            <div className="flex h-16 items-center px-4">
+
+                {/* Header Name */}
+                <div className="flex items-center gap-2">
+                    <div className="flex items-center gap-1 font-semibold text-xl">
+                        <span className="text-primary">StockSage-AI</span>
+                    </div>
+                </div>
+
+                {/* Header top options */}
+                <div className="ml-auto flex items-center gap-4">
+                    <div>
+                        Sign In
+                    </div>
+                </div>
+            </div>
+        </header>
+    )
+}
