+++ frontend/src/components/Sidabar.tsx
@@ -1,0 +1,9 @@
+
+
+export default function Sidebar() {
+    return (
+        <div className="fixed inset-y-0 left-0 z-50 flex w-64 flex-col bg-card transition-transform duration-300 ease-in-out lg:static lg:translate-x-0">
+            Sidebar
+        </div>
+    );
+}
